Split up hasher.
