+++ crates/hasher/src/lib.rs
@@ -1,4 +1,6 @@
 mod hasher;
+mod helpers;
 
-pub use hasher::{create_sha256, Hasher};
+pub use hasher::TargetHasher;
+pub use helpers::*;
 pub use sha2::{Digest, Sha256};
