+++ crates/plugin-node/src/hasher.rs
@@ -1,28 +1,11 @@
-use moon_hasher::{create_sha256, Digest, Hasher, Sha256};
+use moon_hasher::{hash_btree, Digest, Hasher, Sha256};
 use moon_lang_node::{package::PackageJson, tsconfig::TsConfigJson};
-use moon_task::Task;
-use moon_utils::path;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 
 #[derive(Default, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct NodeTargetHasher {
-    // Task `command`
-    command: String,
-
-    // Task `args`
-    args: Vec<String>,
-
-    // Task `deps`
-    deps: Vec<String>,
-
-    // Environment variables
-    env_vars: BTreeMap<String, String>,
-
-    // Input files and globs mapped to a unique hash
-    input_hashes: BTreeMap<String, String>,
-
     // Node.js version
     node_version: String,
 
@@ -34,12 +17,6 @@ pub struct NodeTargetHasher {
 
     // `package.json` `peerDependencies`
     package_peer_dependencies: BTreeMap<String, String>,
-
-    // `project.yml` `dependsOn`
-    project_deps: Vec<String>,
-
-    // Task `target`
-    target: String,
 
     // `tsconfig.json` `compilerOptions`
     tsconfig_compiler_options: BTreeMap<String, String>,
@@ -58,28 +35,6 @@ impl NodeTargetHasher {
         }
     }
 
-    pub fn hash_args(&mut self, passthrough_args: &[String]) {
-        if !passthrough_args.is_empty() {
-            for arg in passthrough_args {
-                self.args.push(arg.clone());
-            }
-
-            // Sort vectors to be deterministic
-            self.args.sort();
-        }
-    }
-
-    /// Hash a mapping of input file paths to unique file hashes.
-    /// File paths *must* be relative from the workspace root.
-    pub fn hash_inputs(&mut self, inputs: BTreeMap<String, String>) {
-        for (file, hash) in inputs {
-            // Standardize on `/` separators so that the hash is
-            // the same between windows and nix machines.
-            self.input_hashes
-                .insert(path::standardize_separators(file), hash);
-        }
-    }
-
     /// Hash `package.json` dependencies as version changes should bust the cache.
     pub fn hash_package_json(&mut self, package: &PackageJson) {
         if let Some(deps) = &package.dependencies {
@@ -93,23 +48,6 @@ impl NodeTargetHasher {
         if let Some(peer_deps) = &package.peer_dependencies {
             self.package_peer_dependencies.extend(peer_deps.clone());
         }
-    }
-
-    /// Hash `dependsOn` from the owning project.
-    pub fn hash_project_deps(&mut self, deps: Vec<String>) {
-        self.project_deps = deps; // Sorted
-    }
-
-    /// Hash `args`, `inputs`, `deps`, and `env` vars from a task.
-    pub fn hash_task(&mut self, task: &Task) {
-        self.command = task.command.clone();
-        self.args = task.args.clone();
-        self.deps = task.deps.clone();
-        self.target = task.target.clone();
-
-        // Sort vectors to be deterministic
-        self.args.sort();
-        self.deps.sort();
     }
 
     /// Hash `tsconfig.json` compiler options that may alter compiled/generated output.
@@ -136,45 +74,14 @@ impl NodeTargetHasher {
 }
 
 impl Hasher for NodeTargetHasher {
-    fn to_hash(&self) -> String {
-        let mut sha = create_sha256();
-
-        let hash_btree = |tree: &BTreeMap<String, String>, hasher: &mut Sha256| {
-            for (k, v) in tree {
-                hasher.update(k.as_bytes());
-                hasher.update(v.as_bytes());
-            }
-        };
-
-        let hash_vec = |list: &Vec<String>, hasher: &mut Sha256| {
-            for v in list {
-                hasher.update(v.as_bytes());
-            }
-        };
-
-        // Order is important! Do not move things around as it will
-        // change the hash and break deterministic builds!
-        // Adding/removing is ok though.
+    fn hash(&self, sha: &mut Sha256) {
         sha.update(self.version.as_bytes());
         sha.update(self.node_version.as_bytes());
 
-        // Task
-        sha.update(self.command.as_bytes());
-        hash_vec(&self.args, &mut sha);
-        hash_vec(&self.deps, &mut sha);
-        hash_btree(&self.env_vars, &mut sha);
-        hash_btree(&self.input_hashes, &mut sha);
-
-        // Deps
-        hash_vec(&self.project_deps, &mut sha);
-        hash_btree(&self.package_dependencies, &mut sha);
-        hash_btree(&self.package_dev_dependencies, &mut sha);
-        hash_btree(&self.package_peer_dependencies, &mut sha);
-
-        // Config
-        hash_btree(&self.tsconfig_compiler_options, &mut sha);
-
-        format!("{:x}", sha.finalize())
+        hash_btree(&self.package_dependencies, sha);
+        hash_btree(&self.package_dev_dependencies, sha);
+        hash_btree(&self.package_peer_dependencies, sha);
+        hash_btree(&self.tsconfig_compiler_options, sha);
     }
 }
